\name{binMixtC-package}
\alias{binMixtC-package}
\alias{binMixtC}
\docType{package}
\title{
\packageTitle{binMixtC}
}
\description{
\packageDescription{binMixtC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{binMixtC}
\packageIndices{binMixtC}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{binMixtC}

Maintainer: \packageMaintainer{binMixtC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
